LayoutTests/platform/mac-gpup/scrollingcoordinator/mac/latching/horizontal-overflow-in-vertical-overflow-expected.txt

Scrolls vertically
Scrolls horizontally
Initial state
PASS outerScroller.scrollTop is 0
PASS innerScroller.scrollTop is 0
After scrolll
FAIL outerScroller.scrollTop should be 210. Was 0.
PASS innerScroller.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
